Time: 10 ms (45.02%) | Memory: 22.8 MB (96.09%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool check(int &x,vector<int>&p,int &h)
     {
-       double cnt=0,n=p.size(),i=0;
+       long long cnt=0,n=p.size(),i=0;
        while(i<n)
        {
         if(p[i]%x!=0)cnt++;
